requirements.txt
python-telegram-bot==21.10
apscheduler==3.11.2
requests==2.32.3
